Add GUI component `GuiMenuFromEnum`
I have done enough GUI in the editor to start to see the common patterns
followed by different features. One of those patterns is having a
draggable menu, with a series of options that can be selected and
based on the selection, a different edit mode is applied when
interacting with the map. Thats the purpose of `GuiMenuFromEnum`. It
works surprisingly well.