Implemented a unified renderer...
Until now each renderer was basically batched and flushed, meaning that you couldnt batch different types of constructs (mixing renderers) and flushin whenever there was a need to us a different renderer. With the new `Renderer` every item is rendered through the renderer and a series of batches as well as their order is recorded, meaning that everything can be flushed together to the screen at once. Also simplifes using the rendering code a lot and will allow to remove the extra `draw` step in the immediate mode gui framework, as soon as I get to it anyway.