Only load application code from application files when they are the root file
It's not really necessary, but in the case of app 1 and 3, they are kind of related applications and so they share some code. For that reason, on importing the file from one to the other, it will also interpret the application for both apps at compile time meaning there will be 2 entry points and other issues. With this changes only the root file will actually have an actual implementation of Application interface.