add minigame "tap tap go"...
The first implementation of a minigame in which you jump through platforms by timing the jump charge.
Also includes some changes to the editor such as the ability to actually add and remove level boxes and some other extras.
Its built on top of app_001 and app_003 so contains a lot of redundant code.