fix regression on zig version 0.12
Starting on 0.12, there is some language level changes related to namespace tpye equivalence: https://ziglang.org/download/0.12.0/release-notes.html#Namespace-Type-Equivalence
The ECS implementation relied on the previous rules to work (kind of hacky, but it is defined behaviour). Since the rules changed slightly it broke the ecs for a bit (resulting in some funny bugs). These changes fixes the situation and the previous behaviour is achieved again.