render the game to a smaller render target and upscale it before blitting it back to the real pixel buffer...
Up till now I was relying on platform mechanism to upscale it back up, which works mostly fine.
However if I do that, my debug text rendering minimum size is forced to be bigger than it could, which is annoying.
So instead, upscale on the application layer and let me address the pixel buffer pixels directly.